Refactor CLI interface for subcommand support
This allows us to do the following:

spotifyutils config --help
spotifyutils playlist --help
